fix initial width on SideNav example story, reorganize props and examples story files, tweak SideNav example story CSS
