+++ src/layout/Panel/props/props.story.tsx
@@ -1,0 +1,10 @@
+import { Panel } from '../Panel';
+
+export default {
+  title: 'layout/Panel/props',
+  component: Panel,
+};
+
+export { Active } from './active.story';
+export { Attach } from './attach.story';
+export { Resizable } from './resizable.story';
